🚚 Move date arbitrary into /arbitrary
